+++ index.js
@@ -6,7 +6,10 @@ var Plasma = module.exports = function(opts){
   this.remoteSubscribers = []
   this.storedChemicals = []
   this.utils = utils
-  this.opts = opts || {missingHandlersChemical: false}
+  this.opts = opts || {
+    missingHandlersChemical: false,
+    throwOnMissingHandler: false
+  }
 }
 
 module.exports.prototype = Object.create(Plasma.prototype)
@@ -90,7 +93,7 @@ module.exports.prototype.store = function (chemical) {
     chemical = {type: chemical}
   }
   this.storedChemicals.push(chemical)
-  return this.emit(chemical)
+  return this._emit(chemical, {skipMissingHandler: true})
 }
 
 module.exports.prototype.storeAndOverride = function (chemical) {
@@ -98,7 +101,7 @@ module.exports.prototype.storeAndOverride = function (chemical) {
     this.trashAll({type: chemical.type})
   }
   this.storedChemicals.push(chemical)
-  return this.emit(chemical)
+  return this._emit(chemical, {skipMissingHandler: true})
 }
 
 module.exports.prototype.has = function (pattern) {
@@ -175,7 +178,13 @@ module.exports.prototype.notifySubscribers = function (chemical) {
   }
 }
 
-module.exports.prototype.emit = function (chemical, callback) {
+/**
+* Private emit function
+* `chemical`: String || Object
+* `options`: {skipMissingHandler: undefined Boolean}
+* `callback`: undefined || Function
+*/
+module.exports.prototype._emit = function (chemical, options, callback) {
   if (typeof chemical == "string") {
     chemical = {type: chemical}
   }
@@ -198,11 +207,67 @@ module.exports.prototype.emit = function (chemical, callback) {
       if (aggregated === true) return true // halt chemical transfer, it has been aggregated
     }
   }
+  
+  if (!options.skipMissingHandler) {
+    if (this.opts.missingHandlersChemical && !hasListeners && chemical.type !== this.opts.missingHandlersChemical) {
+      this.emit({
+        type: this.opts.missingHandlersChemical,
+        chemical: chemical
+      })
+    }
+    if (this.opts.throwOnMissingHandler && !hasListeners) {
+      throw new Error('Missing handler for chemical of type ' + chemical.type)
+    }
+  }
+}
 
-  if (this.opts.missingHandlersChemical && !hasListeners && chemical.type !== this.opts.missingHandlersChemical) {
-    this.emit({
+module.exports.prototype.emit = function (chemical, callback) {
+  return this._emit(chemical, {}, callback)
+}
+
+module.exports.prototype.emitAndCollect = function (chemical, callback) {
+  if (typeof chemical == "string") {
+    chemical = {type: chemical}
+  }
+
+  this.notifySubscribers(chemical)
+  
+  var listenersCount = this.listeners.length
+  var results = []
+  var expectedResultsCount = 0
+  var listeners = []
+  for(var i = 0; i<listenersCount && i<this.listeners.length; i++) {
+    var listener = this.listeners[i]
+    if(this.utils.deepEqual(listener.pattern, chemical)) {
+      if(listener.once) {
+        this.listeners.splice(i, 1);
+        i -= 1;
+        listenersCount -= 1;
+      }
+
+      listeners.push(listener)
+      expectedResultsCount += 1
+    }
+  }
+  if (this.opts.missingHandlersChemical && expectedResultsCount === 0 && chemical.type !== this.opts.missingHandlersChemical) {
+    return this.emit({
       type: this.opts.missingHandlersChemical,
       chemical: chemical
     })
+  }
+  if (this.opts.throwOnMissingHandler && expectedResultsCount === 0) {
+    throw new Error('Missing handler for chemical of type ' + chemical.type)
+  }
+  for (var i = 0; i < listeners.length; i++) {
+    var listener = listeners[i]
+    var aggregated = listener.handler.call(listener.context, chemical, function (err, data) {
+      results.push({err: err, data: data})
+      expectedResultsCount -= 1
+      if (expectedResultsCount === 0) {
+        callback(results)
+      }
+    })
+    // halt chemical transfer, it has been aggregated
+    if (aggregated === true) return callback(results)
   }
 }